+++ bugsnag-android-performance/src/main/kotlin/com/bugsnag/android/performance/internal/AppStartPhase.kt
@@ -1,0 +1,5 @@
+package com.bugsnag.android.performance.internal
+
+enum class AppStartPhase(internal val phaseName: String) {
+    FRAMEWORK("Framework"),
+}