+++ bugsnag-android-performance/src/main/kotlin/com/bugsnag/android/performance/internal/Loopers.kt
@@ -1,0 +1,14 @@
+package com.bugsnag.android.performance.internal
+
+import android.os.Handler
+import android.os.Looper
+
+internal object Loopers {
+    @JvmStatic
+    val main = Looper.getMainLooper()
+
+    val mainHandler = Handler(main)
+
+    fun newMainHandler() = Handler(main)
+    fun newMainHandler(callback: Handler.Callback) = Handler(main, callback)
+}