+++ bugsnag-android-performance/src/main/kotlin/com/bugsnag/android/performance/internal/SpanTracker.kt
@@ -78,6 +78,13 @@ class SpanTracker {
         }
     }
 
+    fun removeAllAssociations(tag: Any?): Collection<SpanImpl> {
+        return lock.write {
+            val associatedSpans = backingStore.remove(tag)
+            associatedSpans.orEmpty().values.map { it.span }
+        }
+    }
+
     /**
      * Mark when a `Span` would have been ended automatically. *If* the associated `Span` is later
      * marked as [leaked](markSpanLeaked) then its `endTime` will be set to the time that this
@@ -125,7 +132,7 @@ class SpanTracker {
     }
 
     /**
-     * End all spans associated with the given `token` or any `subToken`s. This will attempt to
+     * End all spans associated with the given [token] or any `subToken`s. This will attempt to
      * use autoEndTimes where they are available, but will otherwise fallback to using [endTime]
      * for each of the spans that get closed.
      */
@@ -133,6 +140,20 @@ class SpanTracker {
         lock.write {
             val associatedSpans = backingStore[token]
             associatedSpans?.values?.forEach { it.markLeaked(endTime) }
+
+            backingStore.remove(token)
+        }
+    }
+
+    /**
+     * Discard & remove all spans associated with the given [token]. This will result in any
+     * associated spans being discarded, any of their child spans that have already been closed
+     * will have no valid parent span and will also be discarded by the server-side.
+     */
+    fun discardAllSpans(token: Any) {
+        lock.write {
+            val associatedSpans = backingStore[token]
+            associatedSpans?.values?.forEach { it.span.discard() }
 
             backingStore.remove(token)
         }