+++ bugsnag-android-performance/src/test/java/com/bugsnag/android/performance/internal/ConnectivityLegacyTest.kt
@@ -2,6 +2,7 @@
 
 package com.bugsnag.android.performance.internal
 
+import android.content.Intent
 import android.net.ConnectivityManager
 import android.net.NetworkInfo
 import org.junit.Assert.assertEquals
@@ -9,12 +10,20 @@ import org.junit.Test
 import org.junit.runner.RunWith
 import org.junit.runners.Parameterized
 import org.mockito.kotlin.doReturn
+import org.mockito.kotlin.doReturnConsecutively
 import org.mockito.kotlin.mock
 import org.mockito.kotlin.whenever
 
 @RunWith(Parameterized::class)
 class ConnectivityLegacyTest {
     companion object {
+        internal val noNetwork = ConnectivityStatus(
+            false,
+            ConnectionMetering.DISCONNECTED,
+            NetworkType.UNAVAILABLE,
+            null,
+        )
+
         @JvmStatic
         @Parameterized.Parameters
         @Suppress("LongMethod") // naturally long method due to the number of permutations
@@ -99,6 +108,24 @@ class ConnectivityLegacyTest {
         }
 
         val connectivity = ConnectivityLegacy(mock(), connectivityManager, null)
+        val status = connectivity.connectivityStatus
+
+        assertEquals(expected.metering, status.metering)
+        assertEquals(expected.networkType, status.networkType)
+        assertEquals(expected.networkSubType, status.networkSubType)
+    }
+
+    @Test
+    fun testBroadcastNetworkProperties() {
+        val (expected, networkInfo) = testCase
+        val connectivityManager = mock<ConnectivityManager> {
+            whenever(it.activeNetworkInfo) doReturnConsecutively listOf(null, networkInfo)
+        }
+
+        val connectivity = ConnectivityLegacy(mock(), connectivityManager, null)
+        assertEquals(noNetwork, connectivity.connectivityStatus)
+
+        connectivity.onReceive(mock(), Intent())
         val status = connectivity.connectivityStatus
 
         assertEquals(expected.metering, status.metering)