+++ bugsnag-plugin-android-performance-okhttp/src/main/kotlin/com/bugsnag/android/performance/okhttp/BugsnagPerformanceOkhttp.kt
@@ -9,7 +9,6 @@ import okhttp3.Call
 import okhttp3.EventListener
 import okhttp3.Protocol
 import okhttp3.Response
-import okhttp3.internal.headersContentLength
 import java.io.IOException
 import java.util.concurrent.ConcurrentHashMap
 
@@ -32,18 +31,20 @@ class BugsnagPerformanceOkhttp : EventListener() {
             networkSpanOptions,
         )
 
-        val contentLength = call.request().body?.contentLength()
-        if (contentLength != null) {
-            NetworkRequestAttributes.setRequestContentLength(span, contentLength)
-        }
+        if (span != null) {
+            val contentLength = call.request().body?.contentLength()
+            if (contentLength != null) {
+                NetworkRequestAttributes.setRequestContentLength(span, contentLength)
+            }
 
-        spans[call] = span
+            spans[call] = span
+        }
     }
 
     override fun responseHeadersEnd(call: Call, response: Response) {
         val span = spans[call] ?: return
         NetworkRequestAttributes.setResponseCode(span, response.code)
-        val contentLength = response.headersContentLength()
+        val contentLength = response.headers["Content-Length"]?.toLongOrNull() ?: -1L
         if (contentLength != -1L) {
             NetworkRequestAttributes.setResponseContentLength(span, contentLength)
         }